+++ leptos-fluent-macros/src/lib.rs
@@ -75,7 +75,11 @@ use quote::quote;
 ///   macro, which returns [`once_cell:sync::Lazy`]`<[`StaticLoader`]>`.
 /// - **`locales`**: Path to the locales folder, which must contain the translations
 ///   for each language in your application. Is expected to be a path relative from
-///   `Cargo.toml` file.
+///   `Cargo.toml` file, the same used in the [`fluent_templates::static_loader!`]
+///   macro.
+/// - **`core_locales`**: Path to the core locales file, which must contain a shared
+///   translation for all languages. Is expected to be a path relative from `Cargo.toml`,
+///   the same used in the [`fluent_templates::static_loader!`] macro.
 /// - **`check_translations`** (experimental): Path to the files to check if all
 ///    translations are being used and their placeholders are correct. Is expected
 ///   to be a glob pattern relative from `Cargo.toml` file. Tipically, you should use
@@ -190,14 +194,18 @@ pub fn leptos_fluent(
         set_language_to_cookie_bool,
         set_language_to_cookie_expr,
         fluent_file_paths,
+        core_locales_path,
     } = syn::parse_macro_input!(input as I18nLoader);
 
     let n_languages = languages.len();
     let languages_quote = build_languages_quote(&languages);
 
     // files tracker
-    let files_tracker_quote =
-        build_files_tracker_quote(&fluent_file_paths, &languages_path);
+    let files_tracker_quote = build_files_tracker_quote(
+        &fluent_file_paths,
+        &languages_path,
+        &core_locales_path,
+    );
 
     #[cfg(not(feature = "ssr"))]
     let sync_html_tag_lang_quote = {