+++ leptos-fluent/src/lib.rs
@@ -14,7 +14,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! leptos-fluent = "0.0.33"
+//! leptos-fluent = "0.0.34"
 //! fluent-templates = "0.9"
 //!
 //! [features]
@@ -141,8 +141,8 @@
 //!         <p>
 //!             <span>{move || tr!("hello-world")}</span>
 //!             <span>{move_tr!("hello-args", {
-//!                 "arg1" => "value1",
-//!                 "arg2" => "value2",
+//!                 "arg1" => "foo",
+//!                 "arg2" => "bar",
 //!             })}</span>
 //!         </p>
 //!     }